app/requirements.txt
bcrypt==3.2.0
bidict==0.22.0
blinker==1.4
Bootstrap-Flask==2.0.2
cachelib==0.6.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
commonmark==0.9.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.0
Flask==2.1.1
Flask-Assets==2.0
Flask-Bcrypt==1.0.1
Flask-Limiter==2.4.5.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Minify==0.39
Flask-SeaSurf==1.1.1
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==2.5.1
flask-talisman==1.0.0
Flask-WTF==1.0.1
greenlet==1.1.2
htmlmin==0.1.12
idna==3.3
is-safe-url==1.0
itsdangerous==2.1.2
Jinja2==3.1.1
jsmin==3.0.1
lesscpy==0.15.0
limits==2.6.1
MarkupSafe==2.1.1
packaging==21.3
ply==3.11
pycparser==2.21
Pygments==2.12.0
PyJWT==2.3.0
pyparsing==3.0.8
pyScss==1.4.0
python-dotenv==0.19.2
python-engineio==4.3.4
python-socketio==5.7.2
rcssmin==1.1.0
rich==12.3.0
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.2.0
uWSGI==2.0.21
validate-email-address==1
webassets==2.0
Werkzeug==2.1.1
wrapt==1.14.0
WTForms==3.0.1
xxhash==3.0.0
